+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG.md
 
+## [0.1.4] 2025-06-27
+ - Ensured that the refresh rate can be set from the config.
+ - Change options now uses Picker rather than editing the json from nvim.
+
 ## [0.1.3] 2025-06-20
  - Made Aria2TUI class which is called to run the appliction.